% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.log2CPM.R
\name{normalize.log2CPM}
\alias{normalize.log2CPM}
\title{Normalize Data}
\usage{
normalize.log2CPM(X, total = 1e+06, prior = 1)
}
\arguments{
\item{X}{raw data matrix (i.e., gene expression dataset): rows are features/genes; columns are samples.}

\item{total}{total count within each sample}

\item{prior}{offset integer added prior log transformation}
}
\value{
CPM normalized data matrix
}
\description{
Normalize Data
}
\examples{
# load the example dataset provided and run:
nX <- normalize.log2CPM(X)

}
