% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPmatch.R
\name{NPmatch}
\alias{NPmatch}
\title{Batch correction with NPmatch}
\usage{
NPmatch(
  X,
  y,
  dist.method = "cor",
  center.x = TRUE,
  center.m = TRUE,
  knn = 1,
  sdtop = 2000,
  return.B = FALSE,
  use.design = TRUE,
  use.cov = FALSE
)
}
\arguments{
\item{X}{Normalized data matrix: features/genes in rows; samples in columns}

\item{y}{Factor vector indicating batch for each sample}

\item{dist.method}{Distance metric to use for matching ('cor' or 'euclidean')}

\item{center.x}{Logical for whether to center gene expression by row means}

\item{center.m}{Logical for whether to center expression by batch means}

\item{knn}{k nearest neighbor to be determined for each sample in each group}

\item{sdtop}{Number of top variable genes to use}

\item{return.B}{Logical for whether returing the full set of pairs along with the corrected data matrix}

\item{use.design}{Logical for whether using Limma's design matrix}

\item{use.cov}{Logical for whether using a scaled model matrix of the full pairs}
}
\description{
Batch correction with NPmatch
}
\examples{
# Load the example dataset provided and run:
nX <- normalize.log2CPM(X)
cX <- NPmatch(nX)

}
